+++ src/main/java/rbasamoyai/createbigcannons/mixin/ExplosionMixin.java
@@ -2,6 +2,7 @@ package rbasamoyai.createbigcannons.mixin;
 
 import java.util.Optional;
 
+import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -29,12 +30,19 @@ public class ExplosionMixin {
 
 	@WrapOperation(method = "explode",
 		at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/ExplosionDamageCalculator;getBlockExplosionResistance(Lnet/minecraft/world/level/Explosion;Lnet/minecraft/world/level/BlockGetter;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/level/material/FluidState;)Ljava/util/Optional;"))
-	private Optional<Float> createbigcannons$explode(ExplosionDamageCalculator instance, Explosion explosion, BlockGetter reader,
-                                                     BlockPos pos, BlockState state, FluidState fluid, Operation<Optional<Float>> original,
-                                                     @Local(ordinal = 0) float power) {
+	private Optional<Float> createbigcannons$explode$editBlock(ExplosionDamageCalculator instance, Explosion explosion, BlockGetter reader,
+                                                               BlockPos pos, BlockState state, FluidState fluid, Operation<Optional<Float>> original,
+                                                               @Local(ordinal = 0) float power) {
 		if (this instanceof CustomExplosion customExplosion)
 			customExplosion.editBlock(this.level, pos, state, fluid, power);
         return original.call(instance, explosion, reader, pos, state, fluid);
+    }
+
+    @WrapOperation(method = "explode", at = @At(value = "FIELD", target = "Lnet/minecraft/world/level/Explosion;radius:F", ordinal = 1, opcode = Opcodes.GETFIELD))
+    private float createbigcannon$explode$setEntityRadius(Explosion instance, Operation<Float> original) {
+        if (this instanceof CustomExplosion customExplosion)
+            return customExplosion.getEntityRadius();
+        return original.call(instance);
     }
 
 	@WrapOperation(method = "finalizeExplosion",