+++ src/main/java/rbasamoyai/createbigcannons/munitions/big_cannon/smoke_shell/SmokeShellProjectile.java
@@ -29,7 +29,7 @@ public class SmokeShellProjectile extends FuzedBigCannonProjectile {
 	@Override
 	protected void detonate(Position position) {
 		SmokeShellProperties properties = this.getAllProperties();
-		SmokeExplosion explosion = new SmokeExplosion(this.level(), null, position.x(), position.y(), position.z(), 2, Explosion.BlockInteraction.KEEP);
+		SmokeExplosion explosion = new SmokeExplosion(this.level(), null, position.x(), position.y(), position.z(), 0, 2, Explosion.BlockInteraction.KEEP);
 		CreateBigCannons.handleCustomExplosion(this.level(), explosion);
 		SmokeEmitterEntity smoke = CBCEntityTypes.SMOKE_EMITTER.create(this.level());
 		smoke.setPos(new Vec3(position.x(), position.y(), position.z()));