+++ src/main/java/rbasamoyai/createbigcannons/network/ClientboundFluidExplodePacket.java
@@ -15,14 +15,15 @@ import net.minecraft.world.level.material.Fluid;
 import rbasamoyai.createbigcannons.multiloader.EnvExecute;
 import rbasamoyai.createbigcannons.utils.CBCStreamCodecs;
 
-public record ClientboundFluidExplodePacket(double x, double y, double z, float power, List<BlockPos> toBlow, float knockbackX,
-											float knockbackY, float knockbackZ, Fluid fluid) implements RootPacket {
+public record ClientboundFluidExplodePacket(double x, double y, double z, float blockRadius, float entityRadius,
+                                            List<BlockPos> toBlow, float knockbackX, float knockbackY, float knockbackZ, Fluid fluid) implements RootPacket {
 
     public static final StreamCodec<RegistryFriendlyByteBuf, ClientboundFluidExplodePacket> STREAM_CODEC = CBCStreamCodecs.composite(
         ByteBufCodecs.DOUBLE, ClientboundFluidExplodePacket::x,
         ByteBufCodecs.DOUBLE, ClientboundFluidExplodePacket::y,
         ByteBufCodecs.DOUBLE, ClientboundFluidExplodePacket::z,
-        ByteBufCodecs.FLOAT, ClientboundFluidExplodePacket::power,
+        ByteBufCodecs.FLOAT, ClientboundFluidExplodePacket::blockRadius,
+        ByteBufCodecs.FLOAT, ClientboundFluidExplodePacket::entityRadius,
         CatnipStreamCodecBuilders.list(BlockPos.STREAM_CODEC), ClientboundFluidExplodePacket::toBlow,
         ByteBufCodecs.FLOAT, ClientboundFluidExplodePacket::knockbackX,
         ByteBufCodecs.FLOAT, ClientboundFluidExplodePacket::knockbackY,
@@ -36,8 +37,8 @@ public record ClientboundFluidExplodePacket(double x, double y, double z, float 
 	}
 
 	public ClientboundCBCExplodePacket asCBCExplodePacket() {
-		return new ClientboundCBCExplodePacket(this.x, this.y, this.z, this.power, this.toBlow, this.knockbackX, this.knockbackY,
-			this.knockbackZ, ClientboundCBCExplodePacket.ExplosionType.SHRAPNEL);
+		return new ClientboundCBCExplodePacket(this.x, this.y, this.z, this.blockRadius, this.entityRadius, this.toBlow,
+            this.knockbackX, this.knockbackY, this.knockbackZ, ClientboundCBCExplodePacket.ExplosionType.SHRAPNEL);
 	}
 
 }